Move histogram viewer to its own file for clarity.
