+++ src/hubbleds/viewers/__init__.py
@@ -1,3 +1,4 @@
 from .hubble_dotplot import *
+from .hubble_histogram_viewer import *
 from .hubble_scatter_viewer import *
 from .hubble_fit_viewer import *
